lib/iot/src/data_generator/requirements.txt
semantic-version~=2.10.0
wheel~=0.37.1
awscrt~=0.14.6
awsiotsdk~=1.11.8